Add files via upload
In this project, we have detected and identified vehicle number plate that will help in  the detection of car number plates easily. 
This mainly follows Sobel edge detection method and morphological operation for  better approach, and it is simplified to use all the numbers and letters which are in  segmented fashion as it appears in vehicle number plate. Numbers and characters are  recognized using neural networks and bounding box method. The image of the car number plate is captured by the camera of ANPR system and  then it is processed through multiple number of algorithms and then finally we get output of  the detected number plate. These are widely used in highways and airports etc.